refactor: move package to model
