+++ src/main/java/cc/oakk/lotto/model/LottoPrizeProvider.java
@@ -1,4 +1,4 @@
-package cc.oakk.lotto;
+package cc.oakk.lotto.model;
 
 public interface LottoPrizeProvider<T> {
     T getPrizeByRank(Rank rank);
