+++ src/main/java/cc/oakk/lotto/model/LottoResult.java
@@ -1,4 +1,4 @@
-package cc.oakk.lotto;
+package cc.oakk.lotto.model;
 
 import static cc.oakk.lotto.util.ValidationAdapters.throwIfNull;
