+++ src/main/java/cc/oakk/lotto/model/Rank.java
@@ -1,4 +1,4 @@
-package cc.oakk.lotto;
+package cc.oakk.lotto.model;
 
 import java.util.Arrays;
