+++ src/main/java/cc/oakk/lotto/model/SimpleLottoPrizeProvider.java
@@ -1,4 +1,4 @@
-package cc.oakk.lotto;
+package cc.oakk.lotto.model;
 
 import java.util.HashMap;
 import java.util.Map;
