+++ src/test/java/cc/oakk/lotto/model/LottosTest.java
@@ -1,5 +1,7 @@
-package cc.oakk.lotto;
+package cc.oakk.lotto.model;
 
+import cc.oakk.lotto.model.Lotto;
+import cc.oakk.lotto.model.Lottos;
 import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
