+++ src/test/java/cc/oakk/lotto/model/RankTest.java
@@ -1,5 +1,6 @@
-package cc.oakk.lotto;
+package cc.oakk.lotto.model;
 
+import cc.oakk.lotto.model.Rank;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
